requirements.txt
asgiref==3.5.2
Django==4.0.6
django-adminlte-3==0.1.6
django-crispy-forms==1.14.0
django-gravatar2==1.4.4
django-utils-six==2.0
Django-Verify-Email==2.0.3
django-widget-tweaks==1.4.12
Pillow==9.2.0
python-decouple==3.6
sqlparse==0.4.2
tzdata==2022.1
